requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2>=3.1.2
PyJWT==2.8.0
# Database and storage
supabase==2.0.0
psycopg2-binary==2.9.7

# Authentication and security
# Use pre-built cryptography wheels to avoid Rust compilation
cryptography>=41.0.0
python-jose==3.3.0
bcrypt==4.0.1
passlib==1.7.4
python-decouple==3.8

# HTTP requests and API
requests==2.31.0
httpx[http2]==0.24.1
h2==4.1.0
hyperframe==6.0.1
hpack==4.0.0

# WhatsApp/SMS (Twilio)
twilio==8.9.1

# Firebase (for auth tokens)
firebase-admin==6.2.0

# Environment and configuration
python-dotenv==1.0.0

# File handling and utilities
aiofiles==23.2.1
typing-extensions==4.8.0

# Data validation
pydantic[email]==2.5.0

# ASGI server
gunicorn==21.2.0

# PDF generation and manipulation
reportlab==4.0.7
PyMuPDF==1.22.5

# AI and Document Analysis
google-generativeai>=0.3.0
pillow>=10.0.0
python-magic>=0.4.27
PyPDF2>=3.0.0

# Google Cloud Platform
google-cloud-storage>=2.13.0
cloud-sql-python-connector[asyncpg]>=1.4.0
asyncpg>=0.29.0
google-cloud-secret-manager>=2.16.0
SQLAlchemy>=2.0.0

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
